Implemented .JPG Gallery Viewer Activity
Created an activity that allows the user to view a gallery of .JPGs by
swiping left or right. Right now there are no swiping animations, but I
think I've laid the foundation for them to be added later. On the other
hand, I may be using an outdated means of listening for swiping events
and there may be other methods that have built-in animation. It remains
to be seen what our final means of viewing an issue is going to be.